requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aksharamukha==2.1.2
amqp==5.1.1
antlr4-python3-runtime==4.9.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
bardapi==0.1.23a0
beautifulsoup4==4.12.0
billiard==3.6.4.0
boto3==1.26.165
botocore==1.29.165
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
colorama==0.4.6
cryptography==40.0.2
decorator==4.4.2
Deprecated==1.2.13
Django==4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
dnspython==2.3.0
docopt==0.6.2
eventlet==0.33.3
ffmpeg-python==0.2.0
filelock==3.10.7
firebase==4.0.1
firebase-admin==6.1.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
fuzzywuzzy==0.18.0
gdown==4.7.1
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
GoogleBard==1.4.0
googletrans==3.1.0a0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
gTTS==2.3.1
h11==0.9.0
h2==3.2.0
h5py==3.8.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.14.1
hyperframe==5.2.0
icrawler==0.6.6
idna==2.10
imageio==2.27.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.1.0
install==1.3.5
jaconv==0.3.4
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
langcodes==3.3.0
langdetect==1.0.9
language-data==1.1
Levenshtein==0.21.1
libretranslatepy==2.1.1
lit==16.0.2
llvmlite==0.39.1
lxml==4.9.2
marisa-trie==0.7.8
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
more-itertools==9.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mutagen==1.46.0
mysqlclient==2.1.1
networkx==3.0
num2words==0.5.12
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
openai==0.27.2
openai-whisper==20230314
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
pip==23.1.1
proglog==0.1.10
progressbar==2.5
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pykakasi==2.2.1
pyparsing==3.0.9
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-Levenshtein==0.21.1
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.1.1
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
s3transfer==0.6.1
setuptools==67.7.1
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4
sqlparse==0.4.4
super-image==0.1.7
sympy==1.11.1
tiktoken==0.3.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
translate==3.6.1
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
unicodedata2==15.0.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wheel==0.40.0
wincertstore==0.2
wrapt==1.15.0
yarl==1.8.2
zipp==3.15.0
